AudioStore: USE BACKGROUND URLSessionConfiguration, fix progress updating
Downloads were failing the moment the app was no longer in the foreground because it had been using a default URLSessionConfiguration. Somehow this seemed to still work if Watch was using cellular but not wifi?! 🤔

Instead use a background URLSessionConfiguration to allow downloads to continue even when app is put into background.

Set up URLSessionDownloadDelegate didFinishDownloadingTo and didWriteData delegate methods to handle downloads completion and progress updates.

Progress is now calculated with simple division instead of subscribing to Task.progress using Combine.

Remove unused Progress extension.